refactor: consolidate 28 api functions into single router
Vercel Hobby plan caps Serverless Functions at 12 per deployment. After
hero/snake/section/divider/health landed we silently exceeded the cap and
no deploys have succeeded since 2026-04-04 — those endpoints 404'd in
production.

Every handler now lives in src/endpoints/ and is dispatched through a
single api/[name].js router with an explicit allowlist. Brings us down
to 1 Serverless Function; Hobby-friendly again.

Require paths in moved files drop the redundant ../src/ prefix. Test
path for stack handler updated accordingly. All 158 tests pass. Public
URLs unchanged: /api/stats?username=x still reaches the same handler.
