+++ src/endpoints/constellation.js
@@ -1,11 +1,11 @@
-const { renderConstellationCard } = require("../src/cards/constellation");
-const { parseSearchParams, resolveCardOptions } = require("../src/common/options");
+const { renderConstellationCard } = require("../cards/constellation");
+const { parseSearchParams, resolveCardOptions } = require("../common/options");
 const {
   parseColor,
   parseFloatSafe,
   parseIntSafe,
   cacheHeaders,
-} = require("../src/common/utils");
+} = require("../common/utils");
 
 module.exports = async (req, res) => {
   const params = parseSearchParams(req);
