+++ src/endpoints/equalizer.js
@@ -1,6 +1,6 @@
-const { renderEqualizerCard } = require("../src/cards/equalizer");
-const { parseSearchParams, resolveCardOptions } = require("../src/common/options");
-const { parseColor, parseIntSafe, cacheHeaders } = require("../src/common/utils");
+const { renderEqualizerCard } = require("../cards/equalizer");
+const { parseSearchParams, resolveCardOptions } = require("../common/options");
+const { parseColor, parseIntSafe, cacheHeaders } = require("../common/utils");
 
 module.exports = async (req, res) => {
   const params = parseSearchParams(req);
