+++ src/endpoints/heartbeat.js
@@ -1,6 +1,6 @@
-const { renderHeartbeatCard } = require("../src/cards/heartbeat");
-const { parseSearchParams, resolveCardOptions } = require("../src/common/options");
-const { parseColor, parseIntSafe, cacheHeaders } = require("../src/common/utils");
+const { renderHeartbeatCard } = require("../cards/heartbeat");
+const { parseSearchParams, resolveCardOptions } = require("../common/options");
+const { parseColor, parseIntSafe, cacheHeaders } = require("../common/utils");
 
 module.exports = async (req, res) => {
   const params = parseSearchParams(req);
