+++ src/endpoints/stats.js
@@ -1,12 +1,13 @@
-const { fetchReviews } = require("../src/fetchers/reviews");
-const { renderReviewsCard } = require("../src/cards/reviews");
-const { renderError } = require("../src/common/card");
-const { parseSearchParams, resolveCardOptions } = require("../src/common/options");
+const { fetchStats } = require("../fetchers/stats");
+const { renderStatsCard } = require("../cards/stats");
+const { renderError } = require("../common/card");
+const { parseSearchParams, resolveCardOptions } = require("../common/options");
 const {
+  parseArray,
   cacheHeaders,
   errorCacheHeaders,
   classifyError,
-} = require("../src/common/utils");
+} = require("../common/utils");
 
 module.exports = async (req, res) => {
   const params = parseSearchParams(req);
@@ -14,6 +15,8 @@ module.exports = async (req, res) => {
   const { colors, font } = opts;
 
   const username = params.get("username");
+  const hide = parseArray(params.get("hide"));
+  const layout = params.get("layout");
 
   res.setHeader("Content-Type", "image/svg+xml");
   if (themeError) res.setHeader("X-Theme-Error", themeError);
@@ -24,9 +27,11 @@ module.exports = async (req, res) => {
   }
 
   try {
-    // Token pool resolution lives inside fetchReviews via withRotation.
-    const stats = await fetchReviews(username);
-    const svg = renderReviewsCard(stats, opts);
+    // Token pool handling lives in src/common/github-token via withRotation
+    // inside fetchStats — missing/invalid/rate-limited tokens are surfaced
+    // here as thrown errors and rendered as the error card.
+    const stats = await fetchStats(username);
+    const svg = renderStatsCard(stats, { ...opts, hide, layout });
 
     res.setHeader("Cache-Control", cacheHeaders());
     return res.send(svg);
