+++ src/endpoints/tags.js
@@ -1,6 +1,6 @@
-const { renderTagsCard } = require("../src/cards/tags");
-const { parseSearchParams, resolveCardOptions } = require("../src/common/options");
-const { cacheHeaders } = require("../src/common/utils");
+const { renderTagsCard } = require("../cards/tags");
+const { parseSearchParams, resolveCardOptions } = require("../common/options");
+const { cacheHeaders } = require("../common/utils");
 
 module.exports = async (req, res) => {
   const params = parseSearchParams(req);
