+++ src/endpoints/wave.js
@@ -1,17 +1,18 @@
-const { renderDividerCard } = require("../src/cards/divider");
-const { parseSearchParams, resolveCardOptions } = require("../src/common/options");
-const { parseColor, parseIntSafe, cacheHeaders } = require("../src/common/utils");
+const { renderWaveCard } = require("../cards/wave");
+const { parseSearchParams, resolveCardOptions } = require("../common/options");
+const { parseColor, parseIntSafe, cacheHeaders } = require("../common/utils");
 
 module.exports = async (req, res) => {
   const params = parseSearchParams(req);
   const { opts, themeError } = await resolveCardOptions(params);
 
-  const svg = renderDividerCard({
+  const svg = renderWaveCard({
     ...opts,
-    style: params.get("style") || "line",
+    text: params.get("text"),
     color: parseColor(params.get("color")),
     width: parseIntSafe(params.get("width"), 800),
-    height: parseIntSafe(params.get("height"), 30),
+    height: parseIntSafe(params.get("height"), 160),
+    waves: parseIntSafe(params.get("waves"), 3),
   });
 
   res.setHeader("Content-Type", "image/svg+xml");
